experimented with autoencoder training
